Allow an image to span two facing pages, with an optional title.
The image will be split across the two pages.  For best results, an
image should be chosen so that the hard split down the middle does not
split image features that should not be split.

At present, the image is simply printed twice, once on each page, and
clipped appropriately.  This makes the output file larger than it
needs to be.  To conserve space, the image could be split in two, and
the two halves printed separately, but this has not been done yet.

Signed-off-by: Russell Steicke <russells@adelie.cx>