Make the calendars dict use numeric months.
This makes the PostScript code easier to generate, since the dict now
has names like M_2009_10 instead of Oct2009.  This is not important in
US ASCII, but is important when we start using ISO8859-1 or UTF
character sets, as month names in other languages can have characters
that are not legal (I think) in PostScript names.

Also, rename /calendars to /monthCalendars.

Signed-off-by: Russell Steicke <russells@adelie.cx>