Run pcal to parse .calendar files.
The parser for .calendar files has always been incomplete.  Instead of
changing the parser to be a more complete, but possibly still buggy,
.calendar parser, we let pcal do that for us and capture its output.
This means that things like "every monday in all" can now work.

Use the --pcal flag to enable using pcal as the .calendar parser, or
--pcal-planner to enable that and also put these events in the planner
pages.

Beware that the text may be too long to fit in the day boxes on the
planner pages, particularly since when there are multiple events on
one day their text is concatenated, separated by commas.  The text is
truncated when this happens.

Using pcal also means that the #<<short:My birthday>> mechanism is not
available, so there is no way to make the planner text shorter.

Signed-off-by: Russell Steicke <russells@adelie.cx>