Hack around the python installation paths.
On some later python versions (2.6 and onwards on Ubuntu) the paths
used by packages has changed.  I was putting the makediary stuff into
lib/site-python/, but that doesn't work on those versions.

Change the distutils stuff so it works on those versions.

This hack is not ideal, as it detects the python version and changes
behaviour depending on that.  The whole of ./make-deb and ./debian/
need to be reviewed and cleaned up.

Signed-off-by: Russell Steicke <russells@adelie.cx>