Make calendars on the bottom centred around the current month.
Two of the layouts have a row of 12 calendars printed across the
bottom of each opening.  Arrange them so that we print six months up
to and including the current month on the left page, and then six
months following the current month on the right page.

The idea is to make these month calendars more useful, in that their
context is the period immediately before and after the current month,
rather than having them always show the current year, which at the end
of the year will include months from almost a year ago.

Signed-off-by: Russell Steicke <russells@adelie.cx>