Allow the cover page to be an image.
An image to be scaled and maximised to the whole of the front page can
be specified with --cover-page-image=IMAGE.  This image will replace
all of the normal layout on the front page.

This is different from the previous --cover-image=IMAGE option, which
only replaced the default smiley on the cover page.

Because the image replaces the whole of the front page including the
year, it may be a good idea to put the year in the image somewhere.

Signed-off-by: Russell Steicke <russells@adelie.cx>