Allow EPS files to span two pages.
This allows the inclusion of an EPS file, and have it print half on
one (odd) page, and half on the next page.  This might be useful for
maps in particular.

The EPS will be included twice, once for each page, and clipped
appropriately on each page.  There will be no overlap between the two
halves - the EPS is split exactly down the middle.

Signed-off-by: Russell Steicke <russells@adelie.cx>