Adjust the stuff on the inside cover a bit so it fits.
Because our usage message is getting longer, and we are therefore
likely to have long command lines, the debug stuff (--debug-version
output) stuff on the inside cover is getting long.  It can bump up
against the text at the bottom of the page.  So move it up a bit.

Also, make the font on that page slightly smaller.

These measures hide the problem rather than fixing it, but that's
enough for now.

Signed-off-by: Russell Steicke <russells@adelie.cx>