Fix incorrect page sizes in .tex files.
One user reported that the page sizes were incorrect in the .tex
files, and this prevented conversion to pdf and djvu files.  I have
used pdf output and not encountered the issue, but the source was
wrong regardless, so fix that.

Message from Shelagh Manton via github:

Subject: [makediary] Units page (and maybe other info insert pages)
can't be made into a pdf or djvu file. (#1)

I had a look at the units.tex file to see if there was something
strange about it and spotted that the

\ifx \pdfpagewidth \undefined
\else
 \pdfpagewidth=11in    % page width of PDF output
 \pdfpageheight=8.5in  % page height of PDF output
\fi

The height is shorter than the width. When I changed this around the
page was included in the pdf and djvu file with no problem. The
problem only showed up for me because I tried to turn the postscript
file into a pdf and noticed that the resultant file stopped at the
page before the units-ref page.

Signed-off-by: Russell Steicke <russells@adelie.cx>