Add a diary layout without dated pages - a "logbook".
We still print all the front matter (including year calendars) but
just print plain notes pages with a space to write in the dates.  This
simulates a plain work logbook, with a bit more flexibility for note
taking than a diary.  It does miss the ability to schedule things.

The date space is "20__/__/__".  Our descendants in the 22nd century
can change the second digit.

The cover page also changes, so it has spaces to write in the start
and end dates for this diary.

Signed-off-by: Russell Steicke <russells@adelie.cx>