Make a two page image optionally show more than half on each page.
When printing an image over two pages, we have been splitting the
image exactly in half.  This can result in the part of the image in
the middle being split oddly, particularly in the case of images with
text such as maps.

So allow a range between 0.5 and 0.9 (or 50 to 90) to be specified,
meaning that much of the image should be shown on each side.  So a
value of 0.6 means that the left 60% of the image is shown on the left
page, and the right 60% on the right page.  Some of the image will be
duplicated, which is the right thing in some cases.

Signed-off-by: Russell Steicke <russells@adelie.cx>