Command line rationalisation.
Gather all the command line arguments, then process their effect once
we've finished them.  This is to ensure that the effects of each
option are predictable and independent of the command line order.

This change to command line processing is not meant to have any other
effects (except the two options removed below), but is intended to
make some upcoming changes easier.

Also remove the options --day-to-page and --week-to-opening.  Use
--layout=x instead.

Signed-off-by: Russell Steicke <russells@adelie.cx>