Allow specification of a colour for lines and text.
This changes the colour of almost all of the diary images, lines, and
text.  The events are left as black for the events themselves, and
grey for the event warnings.  On the personal information page, the
information itself is black.

Specify a colour with:
    --line-colour=red
    --line-colour=green
    --line-colour=blue
    --line-colour=r,g,b

r, g, and b, are floating point numbers from 0.0 to 1.0.

By default, when a colour is specified, shading is turned off.  It can
be enabled with --shading=yes.

Most of the output is coloured because we set the colour at the start
of each page.  It's possible that some code will unconditionally set
the colour to black and not reset it.  Fix that later.

Signed-off-by: Russell Steicke <russells@adelie.cx>